Adding Supported Application Finder
The FindApps.ps1 script takes a list of applications of interest (i.e.
Quicken, ShadowProtect, eClinicalWorks) and reports back if any of
those applications are found. This is returned along with the
application version, install date, category, and other information.

This is primarily useful when running an audit of a new client so that
common supported applications can be easily found.

The known applications list is saved in XML format. An example
applications.xml file is provided to give an idea of how to format
things. Sometimes effort may have to be given to application names, as
the application name must match (or be a part of) what is shown when
navigating to the “Installed Applications” list.